requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_tflow_select=2.1.0=gpu
absl-py=0.13.0=py37h06a4308_0
aiohttp=3.7.4.post0=py37h7f8727e_2
astor=0.8.1=py37h06a4308_0
async-timeout=3.0.1=py37h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py37h06a4308_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py37h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.9.30=h06a4308_1
cached-property=1.5.2=py_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py37h06a4308_0
cffi=1.14.6=py37h400218f_0
chardet=4.0.0=py37h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.1=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
coverage=5.5=py37h27cfd23_2
cryptography=3.4.8=py37hd23ed53_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cupti=10.0.130=0
cycler=0.10.0=pypi_0
cython=0.29.24=py37h295c915_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.8.1=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h0708190_1
fsspec=2021.8.1=pyhd3eb1b0_0
gast=0.2.2=py37_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.36.1=py37h2157cd5_1
h5py=3.2.1=py37h6c542dc_0
hdf5=1.10.6=hb1b8bf9_0
idna=3.2=pyhd3eb1b0_0
imageio=2.9.0=py_0
importlib-metadata=4.8.1=py37h06a4308_0
intel-openmp=2021.3.0=h06a4308_3350
joblib=1.1.0=pypi_0
jpeg=9d=h36c2ea0_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=h21135ba_2
libprotobuf=3.17.2=h4ff587b_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.0.10=hc3755c2_1005
locket=0.2.1=py37h06a4308_1
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.4=py37h06a4308_0
matplotlib=3.4.3=pypi_0
matplotlib-base=3.4.2=py37hab158f2_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
mock=4.0.3=pyhd3eb1b0_0
multidict=5.1.0=py37h27cfd23_2
munkres=1.1.4=py_0
ncurses=6.2=he6710b0_1
networkx=2.6.3=pyhd3eb1b0_0
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1l=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.0=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pillow=6.2.1=py37h6b7be26_0
pip=21.2.2=py37h06a4308_0
protobuf=3.17.2=py37h295c915_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyjwt=2.1.0=py37h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pysocks=1.7.1=py37_1
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=py37h27cfd23_1
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.15.0=pypi_0
scikit-learn=1.0=pypi_0
scipy=1.7.1=py37h292c36d_2
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tensorboard=1.15.0=pyhb230dea_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=1.15.0=gpu_py37h0f0df58_0
tensorflow-base=1.15.0=gpu_py37h9dcbed7_0
tensorflow-estimator=1.15.1=pyh2649769_0
tensorflow-gpu=1.15.0=h0d30ee6_0
termcolor=1.1.0=py37h06a4308_1
threadpoolctl=3.0.0=pypi_0
tifffile=2021.10.12=pypi_0
tk=8.6.11=h1ccaba5_0
toolz=0.11.1=pyhd3eb1b0_0
torch=1.9.1=pypi_0
torchvision=0.10.1=pypi_0
tornado=6.1=py37h27cfd23_0
tqdm=4.62.3=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
webencodings=0.5.1=py37_1
werkzeug=0.16.1=py_0
wheel=0.37.0=pyhd3eb1b0_1
wrapt=1.12.1=py37h7b6447c_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py37h27cfd23_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=ha95c52a_0
